feat: refactor according to code review comments
